[2 ~ 4단계 방탈출 결제 / 배포] 밍곰(박민선) 미션 제출합니다.  (#313)
* feat: 마이페이지 뷰 코드 업데이트

* feat: 결제 금액을 원시 타입에서 Amount vo로 변경

* feat: 예약 내역 조회시 결제 정보 반환

* chore: `@JoinColumn` name 필드 지정

* chore: `@EnableRetry` 어노테이션 위치 변경 - Application -> Configuration 클래스로 이동

* feat: Swagger ui 추가 및 테마, 시간 생성/삭제 작업을 어드민 권한만 허용

* feat(build.gradle): 로깅 라이브러리 추상화한 `@Slf4j`를 적용하기 위한 lombok 의존성 추가

* feat: logback 설정 파일 추가 및 프로필명 local -> prod 로 변경

* feat(application.yml): open-in-view false

* feat(TossPaymentErrorHandler): 토스 페이먼츠 API 에러 로그 추가

* feat: 로깅을 처리하는 interceptor 정의 - LoggingHandlerInterceptor

* feat: InternalServerErrorException 을 LoggableException으로 변경

* refactor: 어드민 권한 예외 - 401 -> 403으로 변경

* feat: 배포 스크립트 추가

* feat: swagger 문서화 - LoginMember 파라미터 숨기기 및 헤더 필드 지정

* fix: Reservation 필터링 쿼리 파라미터 nullable 로 변경

* refactor: PaymentRestClientConfig 위치 변경 - payment 도메인 패키지 내부로 이동

* fix(ReservationRepository): findAllWithPaymentByMemberId 메서드 - 쿼리 내 memberId 파라미터 누락

* fix(TossPaymentErrorHandler): 로깅 분기문 오류 - else 처리

* refactor(MyReservationResponse): 예약 상태 응답을 나타내는 ReservationStatus 객체 - 내부 클래스로 변환

* refactor: Payment 의 reservationId 필드 삭제 및 조인 테이블 추가

* feat(application.yml): dev 프로필 추가 및 default 프로필을 dev로 변경

* refactor: InternalServerErrorException message 필드 삭제 및 RuntimeException에게 위임

* feat(Amount): 검증 로직 예외를 BadRequestException으로 변경

* feat(ExceptionHandlerAdvice): IllegalArgumentException 핸들링 로직 변경 - 예외 trace 노출되지 않도록 메시지 수정 및 로깅 추가

* feat(ExceptionHandlerAdvice): IllegalStateException 핸들링 로직 및 로깅 추가

* feaTt(Payment): orderId 필드 추가

+++ src/main/java/roomescape/payment/infrastructure/dto/response/ConfirmPaymentResponse.java
@@ -3,7 +3,7 @@ package roomescape.payment.infrastructure.dto.response;
 import roomescape.payment.infrastructure.dto.PaymentFailure;
 
 public record ConfirmPaymentResponse(
-        Integer totalAmount,
+        Long totalAmount,
         String paymentKey,
         PaymentFailure failure
 ) {
